% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.io-package}
\alias{blockr.io}
\alias{blockr.io-package}
\title{blockr.io: Interactive File Import and Export Blocks}
\description{
Extends 'blockr.core' with interactive blocks for reading and writing data files. Supports CSV, Excel, Parquet, RDS, and other formats through a graphical interface without writing code directly. Includes file browser integration and configurable import/export options.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bristolmyerssquibb.github.io/blockr.io/}
  \item Report bugs at \url{https://github.com/BristolMyersSquibb/blockr.io/issues}
}

}
\author{
\strong{Maintainer}: Christoph Sax \email{christoph@cynkra.com} (\href{https://orcid.org/0000-0002-7192-7044}{ORCID})

Authors:
\itemize{
  \item Nicolas Bennett \email{nicolas@cynkra.com}
  \item David Granjon \email{david@cynkra.com}
  \item Mike Page \email{mike@cynkra.com}
}

Other contributors:
\itemize{
  \item  Bristol Myers Squibb [funder]
}

}
\keyword{internal}
