% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanup_uploads}
\alias{cleanup_uploads}
\title{Clean up old uploaded files}
\usage{
cleanup_uploads(upload_dir, max_age_days = 30)
}
\arguments{
\item{upload_dir}{Character. Path to the upload directory.}

\item{max_age_days}{Numeric. Maximum age in days. Files older than this
are removed. Default: 30.}
}
\value{
Invisible NULL.
}
\description{
Removes files older than a given age from an upload directory.
}
\keyword{internal}
