% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-policy.R
\name{collapse_dots}
\alias{collapse_dots}
\title{Lexically collapse \code{.} and \code{..} segments in a path}
\usage{
collapse_dots(path)
}
\description{
Lexically collapse \code{.} and \code{..} segments in a path
}
\keyword{internal}
