% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{new_readcsv_block}
\alias{new_readcsv_block}
\alias{new_writecsv_block}
\title{Upload and read csv files}
\usage{
new_readcsv_block(sep = ",", quote = "\\"", ...)

new_writecsv_block(...)
}
\arguments{
\item{sep}{the field separator character.  Values on each line of the
    file are separated by this character.  If \code{sep = ""} (the
    default for \code{read.table}) the separator is \sQuote{white space},
    that is one or more spaces, tabs, newlines or carriage returns.}

\item{quote}{the set of quoting characters. To disable quoting
    altogether, use \code{quote = ""}.  See \code{\link{scan}} for the
    behaviour on quotes embedded in quotes.  Quoting is only considered
    for columns read as character, which is all of them unless
    \code{colClasses} is specified.}

\item{...}{Forwarded to \code{\link[blockr.core:new_block]{blockr.core::new_block()}}}
}
\description{
This block allows to make avaliable a table from an csv file to a blockr
pipeline.
}
