% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dot_args_names}
\alias{dot_args_names}
\title{Extract names for variadic block arguments}
\usage{
dot_args_names(x)
}
\arguments{
\item{x}{List with names (typically ...args)}
}
\value{
Character vector of names, or NULL if all numeric
}
\description{
Helper function for variadic blocks. Processes ...args names to handle
numeric indices vs named arguments.
}
\keyword{internal}
