% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_url_to_temp}
\alias{download_url_to_temp}
\title{Download URL to temporary file}
\usage{
download_url_to_temp(url)
}
\arguments{
\item{url}{Character. URL to download from.}
}
\value{
Path to temporary file containing downloaded data.
}
\description{
Downloads a file from a URL to a temporary location. Extracts file extension
from URL if available to help with format detection.
}
\keyword{internal}
