% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader_csv.R
\name{get_csv_delimiter}
\alias{get_csv_delimiter}
\title{Get default delimiter for file}
\usage{
get_csv_delimiter(path)
}
\arguments{
\item{path}{Character. Path to file.}
}
\value{
Character. Suggested delimiter based on file extension.
}
\description{
Get default delimiter for file
}
\keyword{internal}
