% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rio_extensions}
\alias{get_rio_extensions}
\title{Get list of file extensions supported by rio}
\usage{
get_rio_extensions()
}
\value{
Character vector of file extensions (without dots)
}
\description{
Returns a comprehensive list of file formats that can be handled by rio::import().
Used for file browser accept parameter and format validation.
}
\keyword{internal}
