% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader_rio.R
\name{is_supported_file}
\alias{is_supported_file}
\title{Check if file is supported}
\usage{
is_supported_file(path)
}
\arguments{
\item{path}{Character. Path to file.}
}
\value{
Logical. TRUE if file format is supported.
}
\description{
Check if file is supported
}
\keyword{internal}
