% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{new_readmulti_block}
\alias{new_readmulti_block}
\alias{new_pick_block}
\title{Read multiple files}
\usage{
new_readmulti_block(
  paths = character(),
  volumes = blockr_option("volumes", c(home = path.expand("~"))),
  ...
)

new_pick_block(which = character(), ...)
}
\arguments{
\item{paths}{File path(s)}

\item{volumes}{Volume path(s)}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}

\item{which}{List selection}
}
\description{
This block allows to read multiple files at once using \code{\link[rio:import]{rio::import()}}.
}
