% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{read_expr_csv}
\alias{read_expr_csv}
\title{Create CSV/TSV/delimited file reading expression}
\usage{
read_expr_csv(path, ...)
}
\arguments{
\item{path}{Character. File path}

\item{...}{Reading parameters: sep, col_names, skip, n_max, quote, encoding}
}
\value{
Expression calling readr::read_csv, readr::read_tsv, or readr::read_delim
}
\description{
Create CSV/TSV/delimited file reading expression
}
\keyword{internal}
