% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{read_expr_excel}
\alias{read_expr_excel}
\title{Create Excel file reading expression}
\usage{
read_expr_excel(path, ...)
}
\arguments{
\item{path}{Character. File path}

\item{...}{Reading parameters: sheet, range, col_names, skip, n_max}
}
\value{
Expression calling readxl::read_excel
}
\description{
Create Excel file reading expression
}
\keyword{internal}
