% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{read_expr_rio}
\alias{read_expr_rio}
\title{Create rio import expression}
\usage{
read_expr_rio(path, ...)
}
\arguments{
\item{path}{Character. File path}

\item{...}{Reading parameters (currently unused)}
}
\value{
Expression calling rio::import
}
\description{
Create rio import expression
}
\keyword{internal}
