% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_names}
\alias{set_names}
\title{Set names helper (base R doesn't export this in older versions)}
\usage{
set_names(x, nm)
}
\description{
Set names helper (base R doesn't export this in older versions)
}
\keyword{internal}
