% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-expr.R
\name{write_expr_csv}
\alias{write_expr_csv}
\title{Build expression to write data to CSV file(s)}
\usage{
write_expr_csv(data_names, path, args = list())
}
\arguments{
\item{data_names}{Character vector of data object names to write}

\item{path}{Character. Full file path for output}

\item{args}{List of write parameters (sep, quote, na, etc.)}
}
\value{
A language object (expression) that writes CSV file(s)
}
\description{
Build expression to write data to CSV file(s)
}
\keyword{internal}
