% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{new_readxlsx_block}
\alias{new_readxlsx_block}
\alias{new_writexlsx_block}
\title{Upload and read Excel}
\usage{
new_readxlsx_block(sheet = NULL, range = NULL, ...)

new_writexlsx_block(...)
}
\arguments{
\item{sheet, range}{See \code{\link[readxl:read_excel]{readxl::read_excel()}}}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}

\item{file}{File name}
}
\description{
This block allows to make avaliable a table from an Excel file to a blockr
pipeline.
}
