% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpt.R
\name{new_readxpt_block}
\alias{new_readxpt_block}
\alias{new_writexpt_block}
\title{Upload and read xpt files}
\usage{
new_readxpt_block(col_select = NULL, skip = 0, n_max = Inf, ...)

new_writexpt_block(...)
}
\arguments{
\item{col_select}{One or more selection expressions, like in
\code{\link[dplyr:select]{dplyr::select()}}. Use \code{c()} or \code{list()} to use more than one expression.
See \code{?dplyr::select} for details on available selection options. Only the
specified columns will be read from \code{data_file}.}

\item{skip}{Number of lines to skip before reading data.}

\item{n_max}{Maximum number of lines to read.}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
This block allows to make avaliable a table from an xpt file to a blockr
pipeline.
}
