% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.io-package}
\alias{blockr.io}
\alias{blockr.io-package}
\title{blockr.io: Reading and writing for blockr}
\description{
Adds blocks for reading and writing various data file types.
}
\author{
\strong{Maintainer}: Nicolas Bennett \email{nicolas@cynkra.com}

}
\keyword{internal}
