% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{read_expr_arrow}
\alias{read_expr_arrow}
\title{Create Arrow file reading expression}
\usage{
read_expr_arrow(path, ...)
}
\arguments{
\item{path}{Character. File path}

\item{...}{Reading parameters (currently unused)}
}
\value{
Expression calling arrow::read_parquet, arrow::read_feather, or arrow::read_ipc_file
}
\description{
Create Arrow file reading expression
}
\keyword{internal}
