% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{read_expr_single}
\alias{read_expr_single}
\title{Create expression for a single file}
\usage{
read_expr_single(path, file_type, ...)
}
\arguments{
\item{path}{Character. Single file path}

\item{file_type}{Character. Type of file}

\item{...}{Parameters for the reader function}
}
\value{
A language object (expression)
}
\description{
Create expression for a single file
}
\keyword{internal}
