% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-expr.R
\name{write_expr_excel}
\alias{write_expr_excel}
\title{Build expression to write data to Excel file with multiple sheets}
\usage{
write_expr_excel(data_names, path)
}
\arguments{
\item{data_names}{Named character vector where names are sheet names}

\item{path}{Character. Full file path for output}
}
\value{
A language object (expression) that writes Excel file
}
\description{
Build expression to write data to Excel file with multiple sheets
}
\keyword{internal}
